+++ changelog.md
@@ -1,7 +1,7 @@
 # Version 0.5.0-alpha - Self Updating.. Update!
 * Rustique now has a self update!! Rustique will update in place. Check out `Rustique help self` to see the commands. (You can check for updates and perform an update)
 * Fixed `list --updates` showing all mods instead of only ones that need updates. This fixes issue #12.
-* `list` now runs sync automatically if the sync file is missing from an installed modpack. This through an error before.
+* `list` now runs sync automatically if the sync file is missing from an installed modpack. This raised an error before.
 * Added `--with-mpk` to the base `Rustique` flags as a shortcut to handling modpack mods. You'll be able to use the base commands, `list`,`update`,`install`, on the packpack mods dir.  `Rustique -w tmmv update -a` -- this will update all the individual mods in the tmmv modpack.
 * You can now install a mod of a specific version with the `install` command, use mod-id@version. Example `Rustique install alchemy@1.6.50`. 
 * Mod backups have been implemented! It's disabled by default, you can turn it on with `Rustique config set --backup-mods true`. You can also choose where they are stored with `Rustique config set --backup-mods-dir /path/to/dir`